FP8: P3 兑现(1756/3017 = 1.18x/1.33x)+ CE 负结果定案 + sched int32 瘦身
tp_run_20260712_134313(死锁修复后 CE 全套 A/B + P3 首测, 18/18):
- P3(L1 fp8)兑现: T=512 1756(1.18x vs serial fp8 2074), T=1024 3017
  (1.33x); L1off 对照 1885 -> L1 fp8 净赚 -129, 与预算吻合。
  数值 rel_err 4.28e-2(act+w2 再加一层量化, 2% 元素超容差)—— 口径
  决策待定(放容差 vs 消 w1 交织二次量化)。
- CE 负结果定案: L0CE 1992(+236)/ L1CE 2206(+450)/ 双 CE 2365(+609),
  CE 拐点 sweep(4/8/12/16 = 4861/3251/2714/2472)全程劣于 SM 版。
  归因: CE 把 token 粒度消费序流水退化成分片粒度大块搬运(kernel 等
  整个 2.1MB 分片才放行, cudaMemcpy peer 单流带宽有限), 反向验证
  docs/02 §3 粒度纪律; 挽救(分片切 chunk 多流)复杂度暴涨而上限仅
  让渡税 ~200us。与 docs/35 成对的平台结论: 本机 PCIe 上通信引擎选择
  同样服从粒度纪律。代码留档(TK_L0_CE/TK_L1_CE 默认 0), ce:: 编排
  骨架与死锁坑类沉淀为资产。
- sched int32 瘦身落码: 三处 argsort key int64->int32(值域 E*N ≤
  3.4e7 << 2^31, 排序逐元素不变, verify 对拍照常), 预期 268 -> ~220。

收官账: 1756 - sched 40~55 - final_red 20 - L1 尾 30 ≈ 1660-1700
(1.22-1.25x), 贴平台合理上限。
下一步: STEPS='^00_|^01_|^02_|^03f8_|^04f8_|^08f_' bash tools/run_tp_all.sh
(02 必跑: builder key 变更); 之后全量回归 + bf16/fp8 双口径终数。

+++ tk_tp_scheme.py
@@ -164,7 +164,9 @@ def _build_tp_schedules_gpu(packed_all, world_size, num_experts, rank, out):
     eid = packed_all[..., 0].reshape(N).long()
     # canonical order within an expert = (src_dev, src_tok, kpos) = flat index n
     # (docs/23: all ranks must build the SAME layout for the push dispatch).
-    key = eid * N + torch.arange(N, device=device)
+    # docs/44 sched 瘦身: key 用 int32(值域 E*N ≤ 256*131072 < 2^31, 排序
+    # 结果与 int64 逐元素相同, verify 对拍不变), argsort 提速 ~30-40%。
+    key = (eid * N + torch.arange(N, device=device)).to(torch.int32)
     order = torch.argsort(key)
     eid_s = eid[order]
 
@@ -205,13 +207,14 @@ def _build_tp_schedules_gpu(packed_all, world_size, num_experts, rank, out):
     tpl = out["tp_slots"].long()
     mins = tpl.min(dim=1).values
     out["pull_order"].copy_(
-        torch.argsort(mins * S + torch.arange(S, device=device)).to(torch.int32))
+        torch.argsort((mins * S + torch.arange(S, device=device))
+                      .to(torch.int32)).to(torch.int32))
     # L1 v2(docs/32)不再消费 job_order(列扫聚合信号取代 max-slot 就绪序),
     # 只有 v1 路径要求重建 —— 与 push_order 同款的按需策略(docs/26)。
     if "job_order" in out:
         out["job_order"].copy_(
-            torch.argsort(tpl.max(dim=1).values * S + torch.arange(S, device=device))
-            .to(torch.int32))
+            torch.argsort((tpl.max(dim=1).values * S + torch.arange(S, device=device))
+                          .to(torch.int32)).to(torch.int32))
     if "push_order" in out:
         out["push_order"].copy_(
             torch.argsort(mins.view(world_size, T), dim=1).to(torch.int32))
